src/requirements.txt
asgiref==3.11.1
cryptography==46.0.1
Django==6.0.3
django-allauth==65.8.1
django-filter==25.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
Markdown==3.10.2
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
